+++ core/shared/src/main/scala/canoe/models/Contact.scala
@@ -12,5 +12,5 @@ package canoe.models
 final case class Contact(phoneNumber: String,
                          firstName: String,
                          lastName: Option[String],
-                         userId: Option[Int],
+                         userId: Option[Long],
                          vcard: Option[String])
